requirements.txt
Flask
pandas
textblob
mysql-connector-python